fix edge case of non existing values in selector (see desc)
When a value was undefined in the list given above, it would, depending on its type, crash the editor. The new way of doing it tries to create a partial instance of the right type for it. This is specific code that works only if well used. This needs even more testing but seems to work without breaking the functionality.